+++ src/main/java/ve/com/cge/appinvoice/accpay/model/Shopping.java
@@ -17,6 +17,7 @@ package ve.com.cge.appinvoice.accpay.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 import javax.persistence.Column;
@@ -69,18 +70,18 @@ public class Shopping implements Serializable {
     private Timestamp date;
     
     @Column(name = "subtotal", nullable = false)
-    private double subTotal;
+    private BigDecimal subTotal;
     
     @Column(name = "amount_tax", nullable = false)
-    private double amountTax;
+    private BigDecimal amountTax;
     
     @Column(name = "total", nullable = false)
-    private double total;
+    private BigDecimal total;
 
     public Shopping() {
     }
 
-    public Shopping(String description, List<ShoppingDetails> shoppingDetails, Supplier supplier, Bank bank, Taxes taxes, Timestamp date, double subTotal, double amountTax, double total) {
+    public Shopping(String description, List<ShoppingDetails> shoppingDetails, Supplier supplier, Bank bank, Taxes taxes, Timestamp date, BigDecimal subTotal, BigDecimal amountTax, BigDecimal total) {
         this.description = description;
         this.shoppingDetails = shoppingDetails;
         this.supplier = supplier;
@@ -92,7 +93,7 @@ public class Shopping implements Serializable {
         this.total = total;
     }
 
-    public Shopping(Long id, String description, List<ShoppingDetails> shoppingDetails, Supplier supplier, Bank bank, Taxes taxes, Timestamp date, double subTotal, double amountTax, double total) {
+    public Shopping(Long id, String description, List<ShoppingDetails> shoppingDetails, Supplier supplier, Bank bank, Taxes taxes, Timestamp date, BigDecimal subTotal, BigDecimal amountTax, BigDecimal total) {
         this.id = id;
         this.description = description;
         this.shoppingDetails = shoppingDetails;
@@ -157,27 +158,27 @@ public class Shopping implements Serializable {
         this.date = date;
     }
 
-    public double getSubTotal() {
+    public BigDecimal getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(double subTotal) {
+    public void setSubTotal(BigDecimal subTotal) {
         this.subTotal = subTotal;
     }
 
-    public double getAmountTax() {
+    public BigDecimal getAmountTax() {
         return amountTax;
     }
 
-    public void setAmountTax(double amountTax) {
+    public void setAmountTax(BigDecimal amountTax) {
         this.amountTax = amountTax;
     }
 
-    public double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(double total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
