+++ src/main/java/ve/com/cge/appinvoice/accrec/model/Billing.java
@@ -17,6 +17,7 @@ package ve.com.cge.appinvoice.accrec.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 import javax.persistence.Column;
@@ -73,18 +74,18 @@ public class Billing implements Serializable {
     private Timestamp date;
     
     @Column(name = "subtotal", nullable = false)
-    private double subTotal;
+    private BigDecimal subTotal;
     
     @Column(name = "amount_tax", nullable = false)
-    private double amountTax;
+    private BigDecimal amountTax;
     
     @Column(name = "total", nullable = false)
-    private double total;
+    private BigDecimal total;
 
     public Billing() {
     }
 
-    public Billing(String description, List<BillingDetails> billingDetails, Customer customer, Seller seller, Bank bank, Taxes taxes, Timestamp date, double subTotal, double amountTax, double total) {
+    public Billing(String description, List<BillingDetails> billingDetails, Customer customer, Seller seller, Bank bank, Taxes taxes, Timestamp date, BigDecimal subTotal, BigDecimal amountTax, BigDecimal total) {
         this.description = description;
         this.billingDetails = billingDetails;
         this.customer = customer;
@@ -97,7 +98,7 @@ public class Billing implements Serializable {
         this.total = total;
     }
 
-    public Billing(Long id, String description, List<BillingDetails> billingDetails, Customer customer, Seller seller, Bank bank, Taxes taxes, Timestamp date, double subTotal, double amountTax, double total) {
+    public Billing(Long id, String description, List<BillingDetails> billingDetails, Customer customer, Seller seller, Bank bank, Taxes taxes, Timestamp date, BigDecimal subTotal, BigDecimal amountTax, BigDecimal total) {
         this.id = id;
         this.description = description;
         this.billingDetails = billingDetails;
@@ -171,27 +172,27 @@ public class Billing implements Serializable {
         this.date = date;
     }
 
-    public double getSubTotal() {
+    public BigDecimal getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(double subTotal) {
+    public void setSubTotal(BigDecimal subTotal) {
         this.subTotal = subTotal;
     }
 
-    public double getAmountTax() {
+    public BigDecimal getAmountTax() {
         return amountTax;
     }
 
-    public void setAmountTax(double amountTax) {
+    public void setAmountTax(BigDecimal amountTax) {
         this.amountTax = amountTax;
     }
 
-    public double getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
 
-    public void setTotal(double total) {
+    public void setTotal(BigDecimal total) {
         this.total = total;
     }
